+++ lib/beyond_api/resources/product_attribute_definitions.rb
@@ -24,19 +24,7 @@ module BeyondApi
     #   @product_attribute_definitions = session.product_attribute_definitions.all(size: 100, page: 0)
     #
     def all(params = {})
-      if params[:paginated] == false
-        result = all_paginated(page: 0, size: 1000)
-
-        (1..result[:page][:total_pages] - 1).each do |page|
-          result[:embedded][:product_attribute_definition].concat(all_paginated(page: page, size: 1000)[:embedded][:product_attribute_definitions])
-        end
-
-        result.is_a?(Hash) ? result.delete(:page) : result.delete_field(:page)
-
-        result
-      else
-        all_paginated(params)
-      end
+      all_results("/product-attribute-definitions", :product_attribute_definitions, params)
     end
 
     #
@@ -114,12 +102,5 @@ module BeyondApi
 
       handle_response(response, status)
     end
-
-    private
-
-      def all_paginated(params = {})
-        response, status = BeyondApi::Request.get(@session, "/product-attribute-definitions", params)
-        handle_response(response, status)
-      end
   end
 end