+++ lib/beyond_api/resources/script_tags.rb
@@ -12,6 +12,7 @@ module BeyondApi
     #   $ curl 'https://api-shop.beyondshop.cloud/api/script-tags' -i -X GET \
     #       -H 'Authorization: Bearer <Access token>'
     #
+    # @option params [Boolean] :paginated
     # @option params [Integer] :size the page size
     # @option params [Integer] :page the page number
     #
@@ -21,9 +22,7 @@ module BeyondApi
     #   @script_tags = session.script_tags.all(size: 20, page: 0)
     #
     def all(params = {})
-      response, status = BeyondApi::Request.get(@session, "/script-tags", params)
-
-      handle_response(response, status)
+      all_results("/script-tags", :script_tags, params)
     end
 
     #